refactored folder structured and changed styles and cards for better UI/UX experience
