+++ src/components/courseCard/courseCard1.js
@@ -1,28 +1,9 @@
-import { Card, CardHeader, Container, IconButton, makeStyles, Typography } from "@material-ui/core";
+import { Card, CardHeader, Container, IconButton, Typography } from "@material-ui/core";
 import React from "react";
+import useStyles from './style';
 import { PlayCircleFilledOutlined } from "@material-ui/icons";
 
-const useStyles = makeStyles((theme) => ({
-  root: {
-    color: "#fff",
-    margin: 0,
-    marginTop: "16px",
-    borderRadius: "8px",
-    padding: 0,
-  },
-  cardHeader: {
-    backgroundColor: "#A60000",
-    color: "#fff",
-  },
-  icon: {
-    color: "#fff",
-    padding: "4px",
-  },
-  title: {
-    fontWeight: 700,
-    fontSize: "12px",
-  },
-}));
+
 
 const CourseCard = () => {
   const classes = useStyles();
@@ -37,7 +18,7 @@ const CourseCard = () => {
             paddingLeft: "8px",
           }}
           title={
-            <Typography variant="h3">
+            <Typography variant="h2">
               Fullstack Web Development Course <br /> by code for cause
             </Typography>
           }
@@ -53,9 +34,9 @@ const CourseCard = () => {
           }}
         />
         <div style={{ backgroundColor: "#000", color: "#fff", padding: "2px 8px" }}>
-          <Typography variant="h6">Chapter 1 : How to start with WD?</Typography>
+          <Typography variant="h5">Chapter 1 : How to start with WD?</Typography>
         </div>
-        <Typography variant="caption" style={{ paddingLeft: "8px", fontStyle: "italic" }}>
+        <Typography className={classes.bottomfont}>
           1hr 45 min learned out of 3 days
         </Typography>
       </Card>
