+++ src/components/wishlistCard/index.js
@@ -4,42 +4,15 @@ import {
   CardHeader,
   Container,
   IconButton,
-  makeStyles,
   Typography,
 } from "@material-ui/core";
 import React from "react";
+import useStyles from './style';
 import ButtonComponent from "../Button/ButtonComponent";
 import { DeleteOutline } from "@material-ui/icons";
 import { useHistory } from "react-router";
 
-const useStyles = makeStyles((theme) => ({
-  root: {
-    color: "#fff",
-    margin: 0,
-    // marginTop: '8px',
-    borderRadius: "8px",
-  },
-  cardHeader: {
-    backgroundColor: "#A60000",
-    color: "#fff",
-  },
-  icon: {
-    color: "#fff",
-    padding: "4px",
-    fontSize: "14px",
-  },
-  title: {
-    fontWeight: 700,
-    fontSize: "12px",
-  },
-  subheader: {
-    fontStyle: "italic",
-    fontSize: "10px",
-    color: "rgb(255, 255, 255, 0.8)",
-    marginTop: "8px",
-    marginBottom: "-8px",
-  },
-}));
+
 
 const WishlistCard = () => {
   const classes = useStyles();
@@ -54,11 +27,15 @@ const WishlistCard = () => {
             backgroundSize: "cover",
           }}
           title={
-            <Typography variant="h3">
+            <Typography variant="h2" style={{ marginBottom:'2px' }}>
               Fullstack Web Development Course <br /> by code for cause
             </Typography>
           }
-          subheader="90hrs of learning with Anuj Garg"
+          subheader={
+            <Typography style={{ fontSize: "14px", marginTop:'5px' }}>
+            90hrs of learning with Anuj Garg
+          </Typography>
+          }
           action={
             <IconButton className={classes.icon}>
               <DeleteOutline />
@@ -72,14 +49,14 @@ const WishlistCard = () => {
           }}
         />
         <CardContent>
-          <Typography>
+          <Typography style={{ fontSize: "14px" }}>
             Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor
             incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam
           </Typography>
-          <Typography variant="caption" style={{ fontStyle: "italic" }}>
+          <Typography variant="caption" style={{ fontStyle: "italic", fontSize: "14px" }}>
             Classes Starting <span style={{ color: "#a60000" }}>16th March Onwards</span>
           </Typography>
-          <Typography variant="h3" color="secondary" style={{ margin: "4px 0px 6px" }}>
+          <Typography variant="h3" color="secondary" style={{ margin: "4px 0px 6px", fontSize: "13.1px" }}>
             ₹ 94
           </Typography>
           <ButtonComponent title="Checkout Now" onClick={() => history.push("/course")} />
